Create index.vue
